requirements.txt
pycryptodome==3.15.0
argparse
xmltodict==0.13.0